# swagger.yaml
openapi: 3.0.0
info:
  title: Your API Documentation
  version: 1.0.0
  description: API documentation for your Express.js application
paths:
  /api/v1/auth/login:
    post:
      summary: User login
      description: Authenticates a user and generates a JWT token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        "400":
          description: Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
  /api/v1/auth/signup:
    post:
      summary: User signup
      description: Creates a new user account
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                fullname:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                isActive:
                  type: boolean
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                  fullname:
                    type: string
                  email:
                    type: string
                  password:
                    type: string
                  role:
                    type: string
                  isActive:
                    type: boolean
        "400":
          description: User already exists or invalid data provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  err:
                    type: string
...